A cursor that iterates over the pixels in a Spot, for a transformed source.
I could not get a super-elegant analytical iterator, so here is
the approach. We favor iterating Z-plane per Z-plane.

First we compute the Zmin and Zmax for a spot (everything is
in pixel units).
Then we compute for each Z plane the intersection of the spot
with the plane, which is a rotated ellipse.
We compute the 2D bounding-box of the ellipse, and create a
cursor over this interval.
The 2D cursor is made to iterate over this interval, but only
stops when it is inside the ellipse.