Pass the scale value from the user dialog to the renderer.
Static value for now. We need this to be visible from the feature
computer too. Might be problematic if the user launches two mastodon
sessions at the same time, but even with that it would not be a great
issue. A proper way to do it would probably be to have some way of
having some kind of model storage from the app.